Close #60
In order not to break the DataBase, now the inserts (or flushes) are
made with 10000 Observations each. This puts less stress in the
DataBase.

Also, related with the issue #61. The Slices can now processed before
the Observations.

The result is a lower memory consumption and faster processing of the
data.